+++ src/Exception/BadRequestError.php
@@ -5,4 +5,4 @@ namespace EyesonTeam\Eyeson\Exception;
 /**
  * eyeson BadRequestError exception
  **/
-class BadRequestError extends \Exception { };
+class BadRequestError extends EyesonApiError { };