+++ src/Exception/EyesonApiError.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace EyesonTeam\Eyeson\Exception;
+
+/**
+ * eyeson EyesonApiError exception
+ **/
+class EyesonApiError extends \Exception { };