+++ src/Exception/ForbiddenError.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace EyesonTeam\Eyeson\Exception;
+
+/**
+ * eyeson ForbiddenError exception
+ **/
+class ForbiddenError extends EyesonApiError { };