+++ src/Exception/GoneError.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace EyesonTeam\Eyeson\Exception;
+
+/**
+ * eyeson GoneError exception
+ **/
+class GoneError extends EyesonApiError { };