+++ src/Exception/NotAcceptableError.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace EyesonTeam\Eyeson\Exception;
+
+/**
+ * eyeson NotAcceptableError exception
+ **/
+class NotAcceptableError extends EyesonApiError { };