+++ src/Exception/NotFoundError.php
@@ -5,4 +5,4 @@ namespace EyesonTeam\Eyeson\Exception;
 /**
  * eyeson NotFoundError exception
  **/
-class NotFoundError extends \Exception { };
+class NotFoundError extends EyesonApiError { };