+++ src/Exception/TimeoutError.php
@@ -5,4 +5,4 @@ namespace EyesonTeam\Eyeson\Exception;
 /**
  * eyeson TimeoutError exception
  **/
-class TimeoutError extends \Exception { };
+class TimeoutError extends EyesonApiError { };