+++ src/Exception/UnknownError.php
@@ -5,4 +5,4 @@ namespace EyesonTeam\Eyeson\Exception;
 /**
  * eyeson UnknownError exception
  **/
-class UnknownError extends \Exception { };
+class UnknownError extends EyesonApiError { };