+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 eyeson-php change log
 
+## v2.1.0 / 2023-02-13
+
+- new: $eyeson->clearWebhook()
+- new: $eyeson->getRecordingById($recordingId)
+- new: $eyeson->createSnapshot($room)
+- new: $eyeson->getSnapshotById($snapshotId)
+- new: EyesonApiError
+    all API request errors are instances of EyesonApiError including
+    getMessage() and getCode()
+    message contains the API response message and code the http status code
+
 ## v2.0.0 / 2023-02-07
 
 - breaking change! $eyeson->getLayout() renamed to $eyeson->layout()